Moved Comment, User, and Tag repositories into the Repositories folder
